pyproject.toml
[project]
name = "parser-kucoin"
version = "2.0.0"
description = ""
authors = [
    {name = "igornet0"}
]
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "undetected-chromedriver (>=3.5.5,<4.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
    "pynput (>=1.8.1,<2.0.0)",
    "kucoin-python (>=1.0.26,<2.0.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "setuptools (>=78.1.0,<79.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "keyboard (>=0.13.5,<0.14.0)",
    "aioconsole (>=0.8.1,<0.9.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "sktime (>=0.37.0,<0.38.0)",
    "torch (>=2.6.0,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "transformers (>=4.51.3,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

